refactoring to avoid circular import dependency b/w profile and cluster pkgs (#43)
* improve updateClusterCommand() and cluster.Update(), refactor to avoid circular dep b/w profile and cluster pkgs

* remove Update(), add comment to Update()
