+++ pkg/common/kvpair.go
@@ -1,0 +1,8 @@
+package common
+
+type KVPair struct {
+	Key   string
+	Value string
+}
+
+type KVPairMap map[string][]KVPair
